# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PharmaSafe Intelligence API",
    "description": "AI-Powered Pharmaceutical Reverse Chain & Medicine Safety Platform Backend Gateway",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/batches": {
      "post": {
        "tags": [
          "Batches & Custody"
        ],
        "summary": "Create Batch",
        "operationId": "create_batch_api_v1_batches_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Batches & Custody"
        ],
        "summary": "List Batches",
        "operationId": "list_batches_api_v1_batches_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BatchStatusEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "medicine_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Medicine Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchResponse"
                  },
                  "title": "Response List Batches Api V1 Batches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id_or_number}/passport": {
      "get": {
        "tags": [
          "Batches & Custody"
        ],
        "summary": "Get Batch Passport",
        "operationId": "get_batch_passport_api_v1_batches__batch_id_or_number__passport_get",
        "parameters": [
          {
            "name": "batch_id_or_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id Or Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPassportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}/recall": {
      "post": {
        "tags": [
          "Batches & Custody"
        ],
        "summary": "Recall Batch",
        "operationId": "recall_batch_api_v1_batches__batch_id__recall_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRecallRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batches/{batch_id}/custody-transfer": {
      "post": {
        "tags": [
          "Batches & Custody"
        ],
        "summary": "Record Custody Transfer",
        "operationId": "record_custody_transfer_api_v1_batches__batch_id__custody_transfer_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustodyTransferCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustodyTransferResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchCreate": {
        "properties": {
          "batch_number": {
            "type": "string",
            "title": "Batch Number",
            "example": "AMX-2026-9901"
          },
          "gtin_barcode": {
            "type": "string",
            "title": "Gtin Barcode",
            "example": "08901234567890"
          },
          "medicine_id": {
            "type": "string",
            "title": "Medicine Id"
          },
          "mfg_date": {
            "type": "string",
            "format": "date",
            "title": "Mfg Date"
          },
          "expiry_date": {
            "type": "string",
            "format": "date",
            "title": "Expiry Date"
          },
          "initial_quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Initial Quantity"
          },
          "unit": {
            "$ref": "#/components/schemas/UnitEnum",
            "default": "BOX"
          },
          "qr_payload": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Payload"
          }
        },
        "type": "object",
        "required": [
          "batch_number",
          "gtin_barcode",
          "medicine_id",
          "mfg_date",
          "expiry_date",
          "initial_quantity"
        ],
        "title": "BatchCreate"
      },
      "BatchPassportResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "batch_number": {
            "type": "string",
            "title": "Batch Number"
          },
          "gtin_barcode": {
            "type": "string",
            "title": "Gtin Barcode"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatusEnum"
          },
          "medicine": {
            "$ref": "#/components/schemas/MedicineResponse"
          },
          "manufacturer_name": {
            "type": "string",
            "title": "Manufacturer Name"
          },
          "mfg_date": {
            "type": "string",
            "format": "date",
            "title": "Mfg Date"
          },
          "expiry_date": {
            "type": "string",
            "format": "date",
            "title": "Expiry Date"
          },
          "initial_quantity": {
            "type": "integer",
            "title": "Initial Quantity"
          },
          "current_quantity": {
            "type": "integer",
            "title": "Current Quantity"
          },
          "unit": {
            "$ref": "#/components/schemas/UnitEnum"
          },
          "current_custodian_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Custodian Name"
          },
          "is_recalled": {
            "type": "boolean",
            "title": "Is Recalled"
          },
          "recall_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall Reason"
          },
          "is_dead_batch": {
            "type": "boolean",
            "title": "Is Dead Batch",
            "default": false
          },
          "destruction_cert_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destruction Cert Hash"
          },
          "custody_history": {
            "items": {
              "$ref": "#/components/schemas/CustodyTransferResponse"
            },
            "type": "array",
            "title": "Custody History",
            "default": []
          },
          "latest_risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Risk Score"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Level"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "batch_number",
          "gtin_barcode",
          "status",
          "medicine",
          "manufacturer_name",
          "mfg_date",
          "expiry_date",
          "initial_quantity",
          "current_quantity",
          "unit",
          "is_recalled",
          "created_at"
        ],
        "title": "BatchPassportResponse"
      },
...
      "BatchResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batch_number": {
            "type": "string",
            "title": "Batch Number"
          },
          "gtin_barcode": {
            "type": "string",
            "title": "Gtin Barcode"
          },
          "medicine_id": {
            "type": "string",
            "title": "Medicine Id"
          },
          "manufacturer_id": {
            "type": "string",
            "title": "Manufacturer Id"
          },
          "mfg_date": {
            "type": "string",
            "format": "date",
            "title": "Mfg Date"
          },
          "expiry_date": {
            "type": "string",
            "format": "date",
            "title": "Expiry Date"
          },
          "initial_quantity": {
            "type": "integer",
            "title": "Initial Quantity"
          },
          "current_quantity": {
            "type": "integer",
            "title": "Current Quantity"
          },
          "unit": {
            "$ref": "#/components/schemas/UnitEnum"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatusEnum"
          },
          "current_custodian_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Custodian Id"
          },
          "is_recalled": {
            "type": "boolean",
            "title": "Is Recalled"
          },
          "recall_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_number",
          "gtin_barcode",
          "medicine_id",
          "manufacturer_id",
          "mfg_date",
          "expiry_date",
          "initial_quantity",
          "current_quantity",
          "unit",
          "status",
          "is_recalled",
          "created_at",
          "updated_at"
        ],
        "title": "BatchResponse"
      },
      "BatchStatusEnum": {
        "type": "string",
        "enum": [
          "MANUFACTURED",
          "IN_DISTRIBUTION",
          "AT_PHARMACY",
          "DISPENSED",
          "EXPIRED",
          "RECALLED",
          "FLAGGED_SUSPICIOUS",
          "RETURN_INITIATED",
          "RETURN_IN_TRANSIT",
          "RECEIVED_AT_DISPOSAL",
          "DESTROYED",
          "DEAD_BATCH"
        ],
        "title": "BatchStatusEnum"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UnitEnum": {
        "type": "string",
        "enum": [
          "TABLET_STRIP",
          "BOTTLE",
          "VIAL",
          "BOX",
          "AMPOULE"
        ],
        "title": "UnitEnum"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}